# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/payment/checkout/one-time": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a Stripe checkout session for one-time payments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "Create a one-time payment checkout session",
                "parameters": [
                    {
                        "description": "One-time checkout details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateOneTimeCheckoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checkout session created",
                        "schema": {
                            "$ref": "#/definitions/handlers.CheckoutResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/payment/checkout/subscription": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a Stripe checkout session for subscription payments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "Create a subscription checkout session",
                "parameters": [
                    {
                        "description": "Checkout session details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateCheckoutSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checkout session created",
                        "schema": {
                            "$ref": "#/definitions/handlers.CheckoutResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/payment/subscription": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns details about the user's current subscription",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "Get subscription details",
                "responses": {
                    "200": {
                        "description": "Subscription details",
                        "schema": {
                            "$ref": "#/definitions/handlers.SubscriptionResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/payment/subscription/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancels the user's subscription at the end of the current billing period",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "Cancel a subscription",
                "responses": {
                    "200": {
                        "description": "Subscription canceled",
                        "schema": {
                            "$ref": "#/definitions/handlers.CancelSubscriptionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/stripe/webhook": {
            "post": {
                "description": "Handles Stripe webhook events for subscription updates, payments, etc.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhook"
                ],
                "summary": "Process Stripe webhook events",
                "responses": {
                    "200": {
                        "description": "Webhook processed",
                        "schema": {
                            "$ref": "#/definitions/handlers.WebhookResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CancelSubscriptionResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Subscription will be canceled at the end of the current billing period"
                },
                "subscription": {
                    "$ref": "#/definitions/handlers.SubscriptionDetails"
                }
            }
        },
        "handlers.CheckoutResponse": {
            "type": "object",
            "properties": {
                "sessionId": {
                    "type": "string",
                    "example": "cs_test_a1b2c3d4e5f6g7h8i9j0"
                },
                "url": {
                    "type": "string",
                    "example": "https://checkout.stripe.com/pay/cs_test_a1b2c3d4e5f6g7h8i9j0"
                }
            }
        },
        "handlers.CreateCheckoutSessionRequest": {
            "type": "object",
            "required": [
                "cancel_url",
                "plan_id",
                "success_url"
            ],
            "properties": {
                "cancel_url": {
                    "type": "string",
                    "example": "https://yourapp.com/cancel"
                },
                "plan_id": {
                    "type": "string",
                    "example": "price_1Oxy3JExamplePriceID"
                },
                "success_url": {
                    "type": "string",
                    "example": "https://yourapp.com/success?session_id={CHECKOUT_SESSION_ID}"
                }
            }
        },
        "handlers.CreateOneTimeCheckoutRequest": {
            "type": "object",
            "required": [
                "amount",
                "cancel_url",
                "currency",
                "product_name",
                "success_url"
            ],
            "properties": {
                "amount": {
                    "description": "Amount in cents, e.g., 2000 = $20.00",
                    "type": "integer",
                    "example": 2000
                },
                "cancel_url": {
                    "type": "string",
                    "example": "https://yourapp.com/cancel"
                },
                "currency": {
                    "type": "string",
                    "example": "usd"
                },
                "product_name": {
                    "type": "string",
                    "example": "Premium Report"
                },
                "success_url": {
                    "type": "string",
                    "example": "https://yourapp.com/success?session_id={CHECKOUT_SESSION_ID}"
                }
            }
        },
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Error message"
                }
            }
        },
...
        "handlers.SubscriptionDetails": {
            "type": "object",
            "properties": {
                "cancel_at_period_end": {
                    "type": "boolean",
                    "example": true
                },
                "current_period_end": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "sub_12345"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                }
            }
        },
        "handlers.SubscriptionResponse": {
            "type": "object",
            "properties": {
                "cancel_at_period_end": {
                    "type": "boolean",
                    "example": false
                },
                "current_period_end": {
                    "type": "string"
                },
                "has_subscription": {
                    "type": "boolean",
                    "example": true
                },
                "plan_id": {
                    "type": "string",
                    "example": "price_1Oxy3JExamplePriceID"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                },
                "subscription_id": {
                    "type": "string",
                    "example": "sub_12345"
                }
            }
        },
...
        "handlers.WebhookResponse": {
            "type": "object",
            "properties": {
                "received": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    }
}